+++ include/chemfiles/formats/TRR.hpp
@@ -11,8 +11,6 @@
 
 #include "chemfiles/files/XDRFile.hpp"
 
-using matrix = float[3][3];
-
 namespace chemfiles {
 class Frame;
 class FormatMetadata;
@@ -28,13 +26,45 @@ class TRRFormat final : public Format {
     size_t nsteps() override;
 
   private:
+    struct FrameHeader {
+        bool use_double; /* Double precision?                                       */
+        int32_t ir_size;     /* Backward compatibility                              */
+        int32_t e_size;      /* Backward compatibility                              */
+        int32_t box_size;    /* Size in Bytes, non zero if a box is present         */
+        int32_t vir_size;    /* Backward compatibility                              */
+        int32_t pres_size;   /* Backward compatibility                              */
+        int32_t top_size;    /* Backward compatibility                              */
+        int32_t sym_size;    /* Backward compatibility                              */
+        int32_t x_size;      /* Size in Bytes, non zero if coordinates are present  */
+        int32_t v_size;      /* Size in Bytes, non zero if velocities are present   */
+        int32_t f_size;      /* Size in Bytes, non zero if forces are present       */
+
+        int32_t natoms;    /* The total number of atoms                           */
+        int32_t step;      /* Current step number                                 */
+        int32_t nre;       /* Backward compatibility                              */
+        double time;   /* Current time (float or double)                          */
+        double lambda; /* Current value of lambda (float or double)               */
+    };
+
+    /// Read header of the Frame at the current position
+    FrameHeader read_frame_header();
+    /// Write header of a Frame
+    void write_frame_header(const FrameHeader& header);
+    /// Determine the number of frames
+    /// and the corresponding offset within the file
+    void determine_frame_offsets();
+
     /// Associated XDR file
     XDRFile file_;
+    /// Offsets within file for fast indexing
+    std::vector<uint64_t> frame_offsets_;
     /// The next step to read
     size_t step_ = 0;
+    /// The number of atoms in the trajectory
+    size_t natoms_ = 0;
 };
 
-template<> const FormatMetadata& format_metadata<TRRFormat>();
+template <> const FormatMetadata& format_metadata<TRRFormat>();
 
 } // namespace chemfiles