+++ src/FormatFactory.cpp
@@ -41,6 +41,7 @@
 #include "chemfiles/formats/TRR.hpp"
 #include "chemfiles/formats/XTC.hpp"
 #include "chemfiles/formats/CIF.hpp"
+#include "chemfiles/formats/DCD.hpp"
 
 #define SENTINEL_INDEX (static_cast<size_t>(-1))
 
@@ -48,10 +49,8 @@ namespace chemfiles {
     class MemoryBuffer;
     class Format;
 
-    extern template class Amber<AMBER_NC_RESTART>;
-    extern template class Amber<AMBER_NC_TRAJECTORY>;
-    extern template class Molfile<DCD>;
     extern template class Molfile<TRJ>;
+    extern template class Molfile<PSF>;
     extern template class Molfile<MOLDEN>;
 }
 using namespace chemfiles;
@@ -63,14 +62,14 @@ static size_t find_by_extension(const std::vector<RegisteredFormat>& formats, st
 
 FormatFactory::FormatFactory() {
     // add formats in alphabetic order
-    this->add_format<Amber<AMBER_NC_RESTART>>();
-    this->add_format<Amber<AMBER_NC_TRAJECTORY>>();
+    this->add_format<AmberRestart>();
+    this->add_format<AmberTrajectory>();
 #ifndef CHFL_DISABLE_GEMMI
     this->add_format<CIFFormat>();
 #endif
     this->add_format<CMLFormat>();
     this->add_format<CSSRFormat>();
-    this->add_format<Molfile<DCD>>();
+    this->add_format<DCDFormat>();
     this->add_format<GROFormat>();
     this->add_format<LAMMPSTrajectoryFormat>();
     this->add_format<LAMMPSDataFormat>();
@@ -79,6 +78,7 @@ FormatFactory::FormatFactory() {
     this->add_format<MOL2Format>();
     this->add_format<Molfile<MOLDEN>>();
     this->add_format<PDBFormat>();
+    this->add_format<Molfile<PSF>>();
     this->add_format<SDFFormat>();
     this->add_format<SMIFormat>();
     this->add_format<TinkerFormat>();