+++ src/formats/SDF.cpp
@@ -65,7 +65,7 @@ void SDFFormat::read_next(Frame& frame) {
 
     frame.reserve(natoms);
     for (size_t i=0; i<natoms; i++) {
-        auto line = file_.readline();
+        line = file_.readline();
 
         if (line.length() < 34) {
             throw format_error(
@@ -116,7 +116,7 @@ void SDFFormat::read_next(Frame& frame) {
     }
 
     for (size_t i=0; i<nbonds; i++) {
-        auto line = file_.readline();
+        line = file_.readline();
         auto atom_1 = parse<size_t>(line.substr(0, 3));
         auto atom_2 = parse<size_t>(line.substr(3, 3));
         auto order = parse<size_t>(line.substr(6, 3));
@@ -148,7 +148,7 @@ void SDFFormat::read_next(Frame& frame) {
     // still be read (until 'M  END' is reached).
     // This loop breaks when the property block ends or returns on an error
     while(!file_.eof()) {
-        auto line = file_.readline();
+        line = file_.readline();
         if (line.empty()) {
             continue;
         } else if (line.substr(0, 4) == "$$$$") {
@@ -165,7 +165,7 @@ void SDFFormat::read_next(Frame& frame) {
     std::string property_name;
     std::string property_value;
     while(!file_.eof()) {
-        auto line = file_.readline();
+        line = file_.readline();
         if (line.empty()) {
             // This breaks a property group - so store now
             if (property_name.empty()) {
@@ -200,8 +200,13 @@ void SDFFormat::write_next(const Frame& frame) {
     auto& positions = frame.positions();
     assert(frame.size() == topology.size());
 
-    // TODO: this can not be more than 80 characters
-    file_.print("{}\n", frame.get<Property::STRING>("name").value_or(""));
+    auto frame_name = frame.get<Property::STRING>("name").value_or("");
+    if (frame_name.size() > 80) {
+        warning("SDF writer", "the frame 'name' property is too long for the SDF format, we truncated it to 80 characters");
+        frame_name = frame_name.substr(0, 80);
+    }
+    file_.print("{}\n", frame_name);
+
     // TODO: this line can contain more data (file creation time and energy in particular)
     file_.print("\n");
     file_.print("created by chemfiles\n");