+++ tests/files/xdr-file.cpp
@@ -2,97 +2,105 @@
 // Copyright (C) Guillaume Fraux and contributors -- BSD license
 
 #include "catch.hpp"
-#include "helpers.hpp"
 #include "chemfiles.hpp"
 #include "chemfiles/files/XDRFile.hpp"
+#include "helpers.hpp"
+#include <vector>
 using namespace chemfiles;
 
-TEST_CASE("XTC files") {
-    SECTION("Read") {
-        XDRFile file(XDRFile::XTC, "data/xtc/ubiquitin.xtc", File::READ);
-
-        CHECK(file.natoms() == 20455);
-        CHECK(file.nframes() == 251);
-        CHECK(file.offset(0) == 0);
-        CHECK(file.offset(10) == 730856);
-
-        CHECK_THROWS_WITH(
-            file.offset(1150),
-            "step 1150 is out of bounds, we have only 251 frames"
-        );
-
-        CHECK_THROWS_WITH(
-            XDRFile(XDRFile::XTC, "not-there.xtc", File::READ),
-            "could not open the file at 'not-there.xtc'"
-        );
-    }
-
-    SECTION("Write") {
-        auto filename = NamedTempPath(".xtc");
-        // just checking constructor/destructor
-        XDRFile file(XDRFile::XTC, filename, File::WRITE);
-    }
-
-    SECTION("Append") {
-        auto filename = NamedTempPath(".xtc");
-        copy_file("data/xtc/ubiquitin.xtc", filename.path());
-        XDRFile file(XDRFile::XTC, filename, File::APPEND);
-
-        CHECK(file.natoms() == 20455);
-        CHECK(file.nframes() == 251);
-        CHECK(file.offset(0) == 0);
-        CHECK(file.offset(10) == 730856);
-
-        auto filename_2 = NamedTempPath(".xtc");
-        file = XDRFile(XDRFile::XTC, filename_2, File::APPEND);
-
-        CHECK(file.natoms() == 0);
-        CHECK(file.nframes() == 0);
-        CHECK_THROWS_WITH(file.offset(0), "step 0 is out of bounds, we have only 0 frames");
-    }
+static void write_xdr_file(XDRFile& file) {
+    file.write_gmx_string("Hello!"); // needs 2B padding
+    const std::vector<float> arrary = {1.234f, -5.123f, 100.232f};
+    file.write_gmx_compressed_floats(arrary, 1000.0);
 }
 
-TEST_CASE("TRR files") {
-    SECTION("Read") {
-        XDRFile file(XDRFile::TRR, "data/trr/water.trr", File::READ);
+TEST_CASE("XDR files") {
+    SECTION("read") {
+        XDRFile file("data/misc/xdr.bin", File::READ);
+        CHECK(file.file_size() == 112);
 
-        CHECK(file.natoms() == 297);
-        CHECK(file.nframes() == 100);
-        CHECK(file.offset(0) == 0);
-        CHECK(file.offset(10) == 36840);
+        // read some big-endian data types
+        CHECK(file.read_single_i32() == -123);
+        CHECK(file.read_single_u32() == 123);
+        CHECK(file.read_single_f64() == 5.678);
+        CHECK(file.read_single_f32() == -4.567f);
 
-        CHECK_THROWS_WITH(
-            file.offset(150),
-            "step 150 is out of bounds, we have only 100 frames"
-        );
+        std::vector<double> darr;
+        darr.resize(3);
+        file.read_f64(darr);
+        const std::vector<double> dexpected = {1.234, -6.234, 105.232};
+        CHECK(darr == dexpected);
 
-        CHECK_THROWS_WITH(
-            XDRFile(XDRFile::TRR, "not-there.trr", File::READ),
-            "could not open the file at 'not-there.trr'"
-        );
+        auto array = std::vector<float>();
+        array.resize(3);
+        file.read_f32(array);
+        const std::vector<float> expected = {1.234f, -5.123f, 100.232f};
+        CHECK(array == expected);
+        array = {0.0, 0.0, 0.0};
+
+        // read XDR and GROMACS specific data types
+        CHECK(file.read_gmx_string() == "Hello!");
+        CHECK(file.read_gmx_compressed_floats(array) == 1000.0f);
+        for (size_t i = 0; i < 3; ++i) {
+            CHECK(approx_eq(array[i], expected[i], 1e-4f));
+        }
     }
 
-    SECTION("Write") {
-        auto filename = NamedTempPath(".trr");
-        // just checking constructor/destructor
-        XDRFile file(XDRFile::TRR, filename, File::WRITE);
+    // clang-format off
+    auto expected = std::vector<uint8_t> {
+        0x00, 0x00, 0x00, 0x07, 0x00, 0x00, 0x00, 0x06,
+        'H', 'e', 'l', 'l', 'o', '!', 0x00, 0x00,
+        0x44, 0x7a, 0x00, 0x00, 0x00, 0x00, 0x04, 0xd2, 0xff, 0xff, 0xeb, 0xfd,
+        0x00, 0x01, 0x87, 0x88, 0x00, 0x00, 0x04, 0xd2, 0xff, 0xff, 0xeb, 0xfd,
+        0x00, 0x01, 0x87, 0x88, 0x00, 0x00, 0x00, 0x48, 0x00, 0x00, 0x00, 0x01,
+        0x42, 0x00, 0x00, 0x00,
+    };
+    // clang-format on
+    auto expected_2 = expected;
+    expected_2.insert(expected_2.end(), expected.begin(), expected.end());
+
+    SECTION("write") {
+        auto filename = NamedTempPath(".bin");
+        {
+            XDRFile file(filename, File::WRITE);
+            write_xdr_file(file);
+            CHECK(file.file_size() == expected.size());
+        }
+
+        auto content = read_binary_file(filename);
+        CHECK(content == expected);
     }
 
-    SECTION("Append") {
-        auto filename = NamedTempPath(".trr");
-        copy_file("data/trr/water.trr", filename.path());
-        XDRFile file(XDRFile::TRR, filename, File::APPEND);
+    SECTION("write and append") {
+        auto filename = NamedTempPath(".bin");
+        {
+            XDRFile file(filename, File::WRITE);
+            write_xdr_file(file);
+            CHECK(file.file_size() == expected.size());
+        }
 
-        CHECK(file.natoms() == 297);
-        CHECK(file.nframes() == 100);
-        CHECK(file.offset(0) == 0);
-        CHECK(file.offset(10) == 36840);
+        auto content = read_binary_file(filename);
+        CHECK(content == expected);
 
-        auto filename_2 = NamedTempPath(".trr");
-        file = XDRFile(XDRFile::TRR, filename_2, File::APPEND);
+        {
+            XDRFile file(filename, File::Mode::APPEND);
+            write_xdr_file(file);
+            CHECK(file.file_size() == expected_2.size());
+        }
 
-        CHECK(file.natoms() == 0);
-        CHECK(file.nframes() == 0);
-        CHECK_THROWS_WITH(file.offset(0), "step 0 is out of bounds, we have only 0 frames");
+        content = read_binary_file(filename);
+        CHECK(content == expected_2);
+    }
+
+    SECTION("append") {
+        auto filename = NamedTempPath(".bin");
+        {
+            XDRFile file(filename, File::WRITE);
+            write_xdr_file(file);
+            CHECK(file.file_size() == expected.size());
+        }
+
+        auto content = read_binary_file(filename);
+        CHECK(content == expected);
     }
 }