+++ tests/formats/sdf.cpp
@@ -1,9 +1,5 @@
 // Chemfiles, a modern library for chemistry file reading and writing
 // Copyright (C) Guillaume Fraux and contributors -- BSD license
-
-#include <streambuf>
-#include <fstream>
-
 #include "catch.hpp"
 #include "helpers.hpp"
 #include "chemfiles.hpp"
@@ -112,7 +108,7 @@ TEST_CASE("Errors in SDF format") {
 
 TEST_CASE("Write files in SDF format") {
     auto tmpfile = NamedTempPath(".sdf");
-    const auto expected_content =
+    const auto EXPECTED_CONTENT =
 R"(
 
 created by chemfiles
@@ -192,6 +188,12 @@ prop1
 1.0 2.0 3.0
 
 $$$$
+abc dfe ghi jkl mno pqr stu vwx yz 123 456 789 ABC DFE GHI JKL MNO PQR STU VWX Y
+
+created by chemfiles
+  0  0  0     0  0  0  0  0  0999 V2000
+M  END
+$$$$
 )";
 
     auto frame = Frame();
@@ -227,33 +229,34 @@ $$$$
     frame.add_bond(9, 10, Bond::AROMATIC);
     frame.add_bond(8, 9, Bond::UNKNOWN);
 
-    frame.set("name", Property("TEST"));
-    frame.set("float property", Property(1.23));
+    frame.set("name", "TEST");
+    frame.set("float property", 1.23);
 
     file.write(frame);
 
     frame.clear_bonds();
     frame.resize(1);
 
-    frame.set("bool property", Property(false));
+    frame.set("bool property", false);
     file.write(frame);
 
-    frame.set("vector property", Property(Vector3D{1.0, 2.0, 3.0}));
+    frame.set("vector property", Vector3D{1.0, 2.0, 3.0});
+    file.write(frame);
+
+    // name is too long for SDF specification
+    frame = Frame();
+    frame.set("name", "abc dfe ghi jkl mno pqr stu vwx yz 123 456 789 ABC DFE GHI JKL MNO PQR STU VWX YZ 123 456 789");
     file.write(frame);
 
     file.close();
 
-    std::ifstream checking(tmpfile);
-    std::string content((std::istreambuf_iterator<char>(checking)),
-                         std::istreambuf_iterator<char>());
-    CHECK(content == expected_content);
+    auto content = read_text_file(tmpfile);
+    CHECK(content == EXPECTED_CONTENT);
 }
 
 TEST_CASE("Read and write files in memory") {
     SECTION("Reading from memory") {
-        std::ifstream checking("data/sdf/kinases.sdf");
-        std::vector<char> content((std::istreambuf_iterator<char>(checking)),
-            std::istreambuf_iterator<char>());
+        auto content = read_text_file("data/sdf/kinases.sdf");
 
         auto file = Trajectory::memory_reader(content.data(), content.size(), "SDF");
         REQUIRE(file.nsteps() == 6);